+++ resources/plugin.yml
@@ -7,4 +7,4 @@ softdepend:
   - "ProtocolSupport"
   - "ViaRewind"
   - "ViaRewind-Legacy-Support"
-main: twolovers.viarewindpotions.Main
+main: dev._2lstudios.viarewindpotions.Main
