+++ src/dev/_2lstudios/viarewindpotions/utils/SplashTranslator.java
@@ -1,4 +1,4 @@
-package twolovers.viarewindpotions.utils;
+package dev._2lstudios.viarewindpotions.utils;
 
 public enum SplashTranslator {
 	NIGHT_VISION(2039713, new TranslationData(8356774, 0, 106), new TranslationData(5, 107, 210)),
